+++ source/core/transform/idwt.cpp
@@ -31,8 +31,8 @@
 #include "utils.hpp"
 
 void idwt_1d_filtr_irrev97_fixed(sprec_t *X, const int32_t left, const int32_t u_i0, const int32_t u_i1) {
-  const auto i0        = static_cast<const int32_t>(u_i0);
-  const auto i1        = static_cast<const int32_t>(u_i1);
+  const auto i0        = static_cast<int32_t>(u_i0);
+  const auto i1        = static_cast<int32_t>(u_i1);
   const int32_t start  = i0 / 2;
   const int32_t stop   = i1 / 2;
   const int32_t offset = left - i0 % 2;
@@ -64,8 +64,8 @@ void idwt_1d_filtr_irrev97_fixed(sprec_t *X, const int32_t left, const int32_t u
 }
 
 void idwt_1d_filtr_rev53_fixed(sprec_t *X, const int32_t left, const int32_t u_i0, const int32_t u_i1) {
-  const auto i0        = static_cast<const int32_t>(u_i0);
-  const auto i1        = static_cast<const int32_t>(u_i1);
+  const auto i0        = static_cast<int32_t>(u_i0);
+  const auto i1        = static_cast<int32_t>(u_i1);
   const int32_t start  = i0 / 2;
   const int32_t stop   = i1 / 2;
   const int32_t offset = left - i0 % 2;
@@ -230,14 +230,14 @@ void idwt_rev_ver_sr_fixed(sprec_t *in, const int32_t u0, const int32_t u1, cons
       for (int32_t col = 0; col < u1 - u0; ++col) {
         int32_t sum = buf[n - 1][col];
         sum += buf[n + 1][col];
-        buf[n][col] -= ((sum + 2) >> 2);
+        buf[n][col] -= static_cast<int16_t>((sum + 2) >> 2);
       }
     }
     for (int32_t n = 0 + offset, i = start; i < stop; ++i, n += 2) {
       for (int32_t col = 0; col < u1 - u0; ++col) {
         int32_t sum = buf[n][col];
         sum += buf[n + 2][col];
-        buf[n + 1][col] += (sum >> 1);
+        buf[n + 1][col] += static_cast<int16_t>(sum >> 1);
       }
     }
